pyproject.toml
[project]
name = "modeljoust"
description = """\
    A tool for computing whether one machine learning model performs \
    statistically significantly better than another.\
"""
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [{name = "Simon Kornblith", email="skornblith@google.com"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]
keywords = []

# pip dependencies of the project
dependencies = [
  "numpy>=1.14",
  "scipy>=0.9",
  "absl-py>=0.8",
  "tabulate>=0.8"
]

# This is set automatically by flit using `modeljoust.__version__`
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/google-research/modeljoust"
repository = "https://github.com/google-research/modeljoust"
# Other: `documentation`, `changelog`

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
    "pytest",
    "pytest-xdist",
    "pylint>=2.6.0",
    "yapf",
]

[build-system]
requires = ["flit_core >=3.5,<4"]
build-backend = "flit_core.buildapi"
